Add clarification comment for command status values
Also extracted test env. for `RejectionBus` tests.
