+++ server/src/test/java/io/spine/server/rejection/given/BareDispatcher.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.rejection.given;
+
+import io.spine.core.RejectionClass;
+import io.spine.core.RejectionEnvelope;
+import io.spine.server.rejection.RejectionDispatcher;
+import io.spine.test.rejection.ProjectRejections;
+
+import java.util.Set;
+
+/**
+ * A simple dispatcher class, which only dispatch and does not have own failure
+ * subscribing methods.
+ */
+public class BareDispatcher implements RejectionDispatcher<String> {
+
+    private boolean dispatchCalled = false;
+
+    @Override
+    public Set<RejectionClass> getMessageClasses() {
+        return RejectionClass.setOf(ProjectRejections.InvalidProjectName.class);
+    }
+
+    @Override
+    public Set<String> dispatch(RejectionEnvelope failure) {
+        dispatchCalled = true;
+        return Identity.of(this);
+    }
+
+    public boolean isDispatchCalled() {
+        return dispatchCalled;
+    }
+}
